Change lat and lng to decimal.

+++ GSS.Model/Search.cs
@@ -16,8 +16,8 @@ namespace GSS.Model
         public int UserId { get; set; }
         public DateTime DateCreated { get; set; }
         public DateTime? DateClosed { get; set; }
-        public double Lat { get; set; }
-        public double Lng { get; set; }
+        public decimal Lat { get; set; }
+        public decimal Lng { get; set; }
 
         public List<Zone> Zones { get; set; }
         public List<Manager> Managers { get; set; }
